+++ CloudHarvestCoreTasks/silos/silos.py
@@ -66,11 +66,20 @@ class BaseSilo:
 
         raise NotImplementedError
 
+    def add_indexes(self, indexes: dict):
+        """
+        Adds indexes to the Silo.
+
+        :param indexes: A dictionary of indexes to add to the Silo.
+        """
+        raise NotImplementedError
+
     def connect(self):
         """
         Connects to the Silo and returns a connection object.
         """
         raise NotImplementedError
+
 
 @register_definition('silo', 'mongo_silo')
 class MongoSilo(BaseSilo):
@@ -98,6 +107,47 @@ class MongoSilo(BaseSilo):
 
         else:
             return False
+
+    def add_indexes(self, indexes: dict):
+        """
+        Create an index in the backend cache.
+
+        Args:
+            indexes (dict): A dictionary containing the indexes to create.
+
+        Returns:
+            None
+        """
+
+        # Get the connection
+        client = self.connect()
+
+        # Identify databases
+        for database in indexes.keys():
+
+            # Identify collections
+            for collection in indexes['harvest'].keys():
+
+                # Identify indexes
+                for index in indexes['harvest'][collection]:
+
+                    # Add single-field indexes defined as a list of strings
+                    if isinstance(index, (str, list)):
+                        client['harvest'][collection].create_index(keys=index)
+                        logger.debug(f'{client.log_prefix}: added index: {database}.{collection}.{str(index)}')
+
+                    # Add complex indexes defined as a dictionary
+                    elif isinstance(index, dict):
+
+                        # pymongo is very picky and demands a list[tuple())
+                        keys = [(i['field'], i.get('sort', 1)) for i in index.get('keys', [])]
+
+                        client['harvest'][collection].create_index(keys=keys, **index['options'])
+
+                        logger.debug(f'{client.log_prefix}: added index: {database}.{collection}.{str(index)}')
+
+                    else:
+                        logger.error(f'unexpected type for index `{index}`: {str(type(index))}')
 
     def connect(self) -> MongoClient:
         """